requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
iso8601==2.1.0
pydantic==2.11.7
pydantic_core==2.33.2
pypika-tortoise==0.6.1
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.2
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
